+++ CHANGELOG.md
@@ -10,6 +10,7 @@ You can also refer to commit logs to get details on what was implemented, fixed 
 - Move `secretInBody` and `customParameters` from `authConfig` to `clientConfig`, where they belong
 - Allow to override default UTF-8 encoding of the _Basic_ authorization header
 - Improvements to embedded authorization
+- Remove `onAuthorize` and `onFailure` callbacks, which have been deprecated with 3.0 (now handled in the callback to `authorize()`)
 
 
 ### 3.0.1